fibonacci ,introduced %ld as placeholder for long int, ISO C90 does not support long long

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -12,18 +12,18 @@
 int main(void)
 {
 	int i = 0;
-	long long int j = 1, k = 2, l = j + k;
+	long int j = 1, k = 2, l = j + k;
 
 	while (i < 98)
 	{
 		if (i == 0)
-			printf("%lli", j);
+			printf("%ld", j);
 		else if (i == 1)
-			printf(", %lli", k);
+			printf(", %ld", k);
 		else
 		{
 			l = j + k;
-			printf(", %lli",l);
+			printf(", %ld",l);
 			j = k;
 			k = l;
 		}
